3491 fix adding a brick
